requirements.txt
numpy>=1.20.0
Pillow>=8.0.0
scipy>=1.7.0
blake3>=0.3.0  # Optional for production
